Allows negative indexing in static list

+++ AlarmClock/Source/Headers/StaticList.hpp
@@ -22,12 +22,12 @@ class StaticList
 		size_t max();
 		T peek();
 		T pop();
-		bool remove(uint8_t index);
+		bool remove(int index);
 		uint8_t size();
 
 		StaticList& operator+=(T value);
-		T operator[](uint8_t index) const;  // Getter
-		T& operator[](uint8_t index);  // Setter
+		T operator[](int index) const;  // Getter
+		T& operator[](int index);  // Setter
 
 	private:
 		uint8_t _size = 0;
